fix java test

+++ binding/java/test/ai/picovoice/leopard/LeopardTest.java
@@ -215,7 +215,10 @@ public class LeopardTest {
 
     @AfterEach
     void tearDown() {
-        leopard.delete();
+        if (leopard != null) {
+            leopard.delete();
+            leopard = null;
+        }
     }
 
     void validateMetadata(
